+++ libraries/map2colour/m2c_colours.h
@@ -1,0 +1,48 @@
+#pragma once
+//
+//    FILE: m2c_colours.h
+//  AUTHOR: Rob Tillaart
+// PURPOSE: Arduino library for mapping a float to colour spectrum
+//     URL: https://github.com/RobTillaart/map2colour
+
+
+//  https://www.w3.org/wiki/CSS/Properties/color/keywords
+#define M2C_BLACK                       0x00000000
+#define M2C_SILVER                      0x00C0C0C0
+#define M2C_GRAY                        0x00808080
+#define M2C_WHITE                       0x00FFFFFF
+#define M2C_MAROON                      0x00800000
+#define M2C_RED                         0x00FF0000
+#define M2C_PURPLE                      0x00800080
+#define M2C_FUCHSIA                     0x00FF00FF
+#define M2C_GREEN                       0x00008000
+#define M2C_LIME                        0x000FF000
+#define M2C_OLIVE                       0x00808000
+#define M2C_YELLOW                      0x00FFFF00
+#define M2C_NAVY                        0x00000080
+#define M2C_BLUE                        0x000000FF
+#define M2C_TEAL                        0x00008080
+#define M2C_AQUA                        0x0000FFFF
+
+
+#define M2C_GRAY_0                      0x00000000
+#define M2C_GRAY_1                      0x00101010
+#define M2C_GRAY_2                      0x00202020
+#define M2C_GRAY_3                      0x00303030
+#define M2C_GRAY_4                      0x00404040
+#define M2C_GRAY_5                      0x00505050
+#define M2C_GRAY_6                      0x00606060
+#define M2C_GRAY_7                      0x00707070
+#define M2C_GRAY_8                      0x00808080
+#define M2C_GRAY_9                      0x00909090
+#define M2C_GRAY_10                     0x00A0A0A0
+#define M2C_GRAY_11                     0x00B0B0B0
+#define M2C_GRAY_12                     0x00C0C0C0
+#define M2C_GRAY_13                     0x00D0D0D0
+#define M2C_GRAY_14                     0x00E0E0E0
+#define M2C_GRAY_15                     0x00F0F0F0
+#define M2C_GRAY_16                     0x00FFFFFF
+
+
+//  -- END OF FILE --
+